Add postgresql and link with app
Add postgres to the application as the database that will be used. 
This involves adding it to docker-compose so that it can easily 
be instantiated. 
Alongside this a number of properties have been added to the
application.properties to override some of the default choices. 
By default this points to localhost but is overridden in the docker-compose
by the application environment to point to the running docker image
in the network.