Add preliquibase and liquibase
Add liquibase to support the versioning of the database.
This is not actually important for this project, but acts as a proof 
of concept for other projects. 
As the schema might not necessarily exist at the start and Liquibase 
does not create them by default. 
Preliquibase was added that creates the required schema if it
does not exist already.