Add hazelcast caching to application
To support caching through the application, this change 
introduces the spring boot caching mechanic. 
This involves including the caching starter spring boot dependancy.
As we have hazelcast included in the project this allows this 
to take control of the default spring boot caching mechanism. 
Caching has also been enabled through the ItemService.